Rename NullablePropertyFactory.CodeGenerator -> NullableProperty
NullablePropertyFactory is now NullableProperty.Factory
